Added NFC transfering support for Android (hopefully it works)
Added an options page because I needed somewhere the put options. The UI is not the most amazing thing in the world.
Had to create an android project to interface with NFC.

Should fix problems with 'Temporary Commit' that apparently I created at some point not sure why but it is there.